Update examples and the readme.
