+++ examples/using-router/components/Header.js
@@ -1,9 +1,31 @@
-import ActiveLink from './ActiveLink'
+import Router from 'next/router'
 
 export default () => (
   <div>
-    <ActiveLink href='/'>Home</ActiveLink>
-    <ActiveLink href='/about'>About</ActiveLink>
-    <ActiveLink href='/error'>Error</ActiveLink>
+    <Link href='/'>Home</Link>
+    <Link href='/about'>About</Link>
+    <Link href='/error'>Error</Link>
   </div>
+)
+
+// typically you want to use `next/link` for this usecase
+// but this example shows how you can also access the router
+// and use it manually
+
+function onClickHandler (href) {
+  return (e) => {
+    e.preventDefault()
+    Router.push(href)
+  }
+}
+
+const Link = ({ children, href }) => (
+  <a href='#' onClick={onClickHandler(href)}>
+    {children}
+    <style jsx>{`
+      a {
+        margin-right: 10px;
+      }
+    `}</style>
+  </a>
 )
