+++ examples/using-with-router/components/Header.js
@@ -1,0 +1,9 @@
+import ActiveLink from './ActiveLink'
+
+export default () => (
+  <div>
+    <ActiveLink href='/'>Home</ActiveLink>
+    <ActiveLink href='/about'>About</ActiveLink>
+    <ActiveLink href='/error'>Error</ActiveLink>
+  </div>
+)
