+++ examples/using-with-router/package.json
@@ -7,6 +7,7 @@
     "start": "next start"
   },
   "dependencies": {
+    "hoist-non-react-statics": "^2.2.2",
     "next": "latest",
     "react": "^15.4.2",
     "react-dom": "^15.4.2"
