+++ examples/using-with-router/pages/about.js
@@ -1,0 +1,8 @@
+import Header from '../components/Header'
+
+export default () => (
+  <div>
+    <Header />
+    <p>This is the about page.</p>
+  </div>
+)
