+++ examples/using-with-router/pages/error.js
@@ -3,16 +3,11 @@ import Header from '../components/Header'
 import Router from 'next/router'
 
 export default class extends Component {
-  static getInitialProps () {
-    console.log(Router.pathname)
-    return {}
-  }
-
   render () {
     return (
       <div>
         <Header />
-        <p>This should not be rendered via SSR</p>
+        <p>This path({Router.pathname}) should not be rendered via SSR</p>
       </div>
     )
   }
