+++ examples/using-with-router/pages/index.js
@@ -1,0 +1,8 @@
+import Header from '../components/Header'
+
+export default () => (
+  <div>
+    <Header />
+    <p>HOME PAGE is here!</p>
+  </div>
+)
